+++ R/tag-grobs.R
@@ -87,59 +87,5 @@ makeContent.EulerTags <- function(x) {
   padding <- eulerr_options()$padding
   padding <- convertHeight(padding, "native", TRUE)
 
-  if (!x$adjust_labels)
-    return(x)
-
-  d <- lapply(x$children, function(z) {
-    lab <- z$children[[1]]
-
-    if (inherits(lab, "null")) {
-      out <- c(x = NA, y = NA, w = NA, h = NA)
-    } else {
-      c(x = convertX(lab$x, "native", TRUE),
-        y = convertY(lab$y, "native", TRUE),
-        w = convertWidth(grobWidth(lab), "native", TRUE),
-        h = convertHeight(grobHeight(lab), "native", TRUE))
-    }
-  })
-
-  d <- as.data.frame(do.call(rbind, d), stringsAsFactors = TRUE)
-
-  nas <- is.na(d$x)
-
-  d <- d[!nas, , drop = FALSE]
-
-  # only do something if there are any labels
-  if (NROW(d) > 0) {
-    # boxes are the labels (in eulerr terminology)
-    boxes <- data.frame(x1 = d$x - 0.5*d$w - padding,
-                        y1 = d$y - 0.5*d$h - padding,
-                        x2 = d$x + 0.5*d$w + padding,
-                        y2 = d$y + 0.5*d$h + padding,
-                        stringsAsFactors = TRUE)
-
-    repel <- repel_boxes(
-      boxes = as.matrix(boxes),
-      xlim = xlim,
-      ylim = ylim,
-      hjust = 0.5,
-      vjust = 0.5,
-      force_push = 1e-6,
-      force_pull = 1e-6,
-      maxiter = 1e5,
-      direction = "both"
-    )
-
-    k <- 1
-
-    for (i in seq_along(x$children)) {
-      if (!inherits(x$children[[i]]$children[[1]], "null")) {
-        x$children[[i]]$children[[1]]$x <- unit(repel[k, 1], "native")
-        x$children[[i]]$children[[1]]$y <- unit(repel[k, 2], "native")
-        k <- k + 1
-      }
-    }
-  }
-
   x
 }