+++ tests/testthat/test-plotting.R
@@ -117,17 +117,3 @@ test_that("plots with euler lists works", {
   dev.off()
   unlink(tmp)
 })
-
-test_that("label repelling functions", {
-  tmp <- tempfile()
-  png(tmp, width = 150, height = 150)
-
-  set.seed(2112)
-  f1 <- euler(c("very long label that has lots of words in it" = 1,
-                "another long, long label that is sure to overlap" = 1,
-                "very long label that has lots of words in it&another long, long label that is sure to overlap" = 10))
-  expect_silent(plot(f1, adjust_labels = TRUE, quantities = TRUE))
-
-  dev.off()
-  unlink(tmp)
-})